requirements.txt
astroid==2.4.2
awesome-slugify==1.6.5
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudscraper==1.2.48
colorama==0.4.4
feedgen==0.9.0
Flask==2.0.1
Flask-Caching==1.9.0
idna==2.10
isort==5.6.4
itsdangerous==2.0.0
Jinja2==3.0.0
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==2.0.0rc2
mccabe==0.6.1
natsort==7.1.1
packaging==20.7
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-redis-rate-limit==0.0.7
python-resumable==0.1.2
pytz==2021.1
redis==3.5.3
regex==2020.11.13
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tabulate==0.8.7
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
ujson==4.0.2
Unidecode==0.4.21
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==2.0.0
wrapt==1.12.1
